requirements.txt
Flask==2.3.2
python-dotenv
cohere
groq==0.28.0
google-api-python-client
beautifulsoup4
pywhatkit
requests
rich
spacy
fuzzywuzzy
googlesearch-python
deep-translator
aiohttp
python-Levenshtein
gunicorn
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.7.0/xx_ent_wiki_sm-3.7.0-py3-none-any.whl
